Separate notificaitons and transactions
